refactor:#2 create TodoTable
