+++ go.mod
@@ -6,5 +6,4 @@ require (
 	github.com/go-ini/ini v1.67.0
 	github.com/google/uuid v1.6.0
 	github.com/lib/pq v1.10.9
-	github.com/stretchr/testify v1.9.0 // indirect
 )
