+++ log.text
@@ -11,3 +11,5 @@
 2024/05/19 13:40:50 main.go:19: this is a TEST
 2024/06/24 09:14:18 base.go:45: pq: password authentication failed for user "ubuntu"
 2024/06/24 09:15:49 main.go:30: sql: no rows in result set
+2024/06/27 04:37:09 base.go:46: pq: password authentication failed for user "ubuntu"
+2024/06/27 04:38:49 main.go:27: sql: no rows in result set
