+++ main.go
@@ -4,15 +4,13 @@ import (
     "main/models"
     
     "fmt"
-    _"log"
+    "log"
 
     )
 
 
 
 func main(){
-    user1 := user.NewUser("A123","Yusaku")
-    fmt.Println(user1)
     
     //log.Println("this is a TEST")
