+++ models/user.go
@@ -1,4 +1,4 @@
-package app
+package models
 
 import(
     "time"
